Fix: Floating point for jury points

+++ PlannerApi/Deserializers/EventDeserializer.cs
@@ -28,7 +28,7 @@ namespace events_planner.Deserializers {
         public string Status { get; set; } = Models.Status.PENDING;
 
         [Range(0, int.MaxValue, ErrorMessage = "The Jury Points must be superior to 0")]
-        public int? JuryPoint { get; set; }
+        public float? JuryPoint { get; set; }
 
         public string Location { get; set; }
